backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
torch>=2.6.0
torchvision>=0.19.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=2.5.0
aiofiles>=23.2.1